requirements/base.txt
Django==1.6.1
South
django-crispy-forms
pytz
django-gravatar2
pillow
fabric
raven
django-countries
django-simple-captcha
beautifulsoup4
django-tastypie
python-mimeparse
python-dateutil
lxml
pyyaml
django-celery-with-redis
django-reversion
bbcode
html2bbcode
html2text
django-grappelli
sorl-thumbnail
psycopg2
markdown
pygments
django-braces
django-bootstrap-pagination==1.5.0

# Searching with elastic
django-haystack==2.1.0
requests==2.2.0
pyelasticsearch==0.6.1
